Creating CNAME file in gh-pages branch!
In this commit, an update to the repository has been made by creating a CNAME file in the gh-pages branch of this repository. In order GitHub pages is able to work, this repository needs a gh-pages branch which the GitHub pages engine will specifically use to render your website source code. The mistake that we were doing was that we were updating the __"Master"__ instead of the __"gh-pages"__ . Funny? LOL! :see_no_evil:  :hear_no_evil:  :speak_no_evil: